requirements.txt
pyyaml
importlib_resources
numpy
pytest
jsonschema
pytest-cov
